mod: 백엔드 디렉토리 구조 변경
