+++ backend/storage/src/tests/fetchFileMetadata.test.mjs
@@ -1,17 +1,18 @@
-// getUserUsage.test.mjs
+// fetchFileMetadata.test.mjs
 import AWSMock from 'aws-sdk-mock';
 import AWS from 'aws-sdk';
-import { handler } from '../service/getUserUsage.mjs';
+import { handler } from '../handlers/fetchFileMetadata.mjs';
 
 // AWS SDK Mock 설정
 AWSMock.setSDKInstance(AWS);
 
-describe('getUserUsage Lambda Function', () => {
+describe('fetchFileMetadata Lambda Function', () => {
   beforeAll(() => {
     // DynamoDB query 메서드를 mock 처리
     AWSMock.mock('DynamoDB.DocumentClient', 'query', (params, callback) => {
-      if (params.ExpressionAttributeValues[':userId'] === 'validUserId') {
-        callback(null, { Items: [{ id: '1', usage: '100MB' }] }); // 정상 쿼리 시
+      if (params.ExpressionAttributeValues[':targetId'] === 'validTargetId' &&
+          params.ExpressionAttributeValues[':targetType'] === 'validTargetType') {
+        callback(null, { Items: [{ id: '1', name: 'Test File' }] }); // 정상 쿼리 시
       } else {
         callback(null, { Items: [] }); // 결과가 없는 경우
       }
@@ -22,34 +23,38 @@ describe('getUserUsage Lambda Function', () => {
     AWSMock.restore('DynamoDB.DocumentClient');
   });
 
-  it('should fetch user usage successfully', async () => {
+  it('should fetch file metadata successfully', async () => {
     const event = {
       queryStringParameters: {
-        userId: 'validUserId',
+        targetId: 'validTargetId',
+        targetType: 'validTargetType',
       },
     };
 
     const response = await handler(event);
 
     expect(response.statusCode).toBe(200);
-    expect(JSON.parse(response.body)).toEqual([{ id: '1', usage: '100MB' }]);
+    expect(JSON.parse(response.body)).toEqual([{ id: '1', name: 'Test File' }]);
   });
 
-  it('should return 400 if userId is missing', async () => {
+  it('should return 400 if targetId or targetType is missing', async () => {
     const event = {
-      queryStringParameters: {},
+      queryStringParameters: {
+        targetId: 'validTargetId',
+      },
     };
 
     const response = await handler(event);
 
     expect(response.statusCode).toBe(400);
-    expect(JSON.parse(response.body).error).toBe('Please provide userId');
+    expect(JSON.parse(response.body).error).toBe('Please provide both targetId and targetType.');
   });
 
-  it('should return 200 with an empty array if no usage data is found', async () => {
+  it('should return 200 with an empty array if no metadata is found', async () => {
     const event = {
       queryStringParameters: {
-        userId: 'invalidUserId',
+        targetId: 'invalidTargetId',
+        targetType: 'invalidTargetType',
       },
     };
 
@@ -66,7 +71,8 @@ describe('getUserUsage Lambda Function', () => {
 
     const event = {
       queryStringParameters: {
-        userId: 'validUserId',
+        targetId: 'validTargetId',
+        targetType: 'validTargetType',
       },
     };
