+++ backend/storage/src/tests/getUserUsage.test.mjs
@@ -1,18 +1,17 @@
-// fetchFileMetadata.test.mjs
+// getUserUsage.test.mjs
 import AWSMock from 'aws-sdk-mock';
 import AWS from 'aws-sdk';
-import { handler } from '../service/fetchFileMetadata.mjs';
+import { handler } from '../handlers/getUserUsage.mjs';
 
 // AWS SDK Mock 설정
 AWSMock.setSDKInstance(AWS);
 
-describe('fetchFileMetadata Lambda Function', () => {
+describe('getUserUsage Lambda Function', () => {
   beforeAll(() => {
     // DynamoDB query 메서드를 mock 처리
     AWSMock.mock('DynamoDB.DocumentClient', 'query', (params, callback) => {
-      if (params.ExpressionAttributeValues[':targetId'] === 'validTargetId' &&
-          params.ExpressionAttributeValues[':targetType'] === 'validTargetType') {
-        callback(null, { Items: [{ id: '1', name: 'Test File' }] }); // 정상 쿼리 시
+      if (params.ExpressionAttributeValues[':userId'] === 'validUserId') {
+        callback(null, { Items: [{ id: '1', usage: '100MB' }] }); // 정상 쿼리 시
       } else {
         callback(null, { Items: [] }); // 결과가 없는 경우
       }
@@ -23,38 +22,34 @@ describe('fetchFileMetadata Lambda Function', () => {
     AWSMock.restore('DynamoDB.DocumentClient');
   });
 
-  it('should fetch file metadata successfully', async () => {
+  it('should fetch user usage successfully', async () => {
     const event = {
       queryStringParameters: {
-        targetId: 'validTargetId',
-        targetType: 'validTargetType',
+        userId: 'validUserId',
       },
     };
 
     const response = await handler(event);
 
     expect(response.statusCode).toBe(200);
-    expect(JSON.parse(response.body)).toEqual([{ id: '1', name: 'Test File' }]);
+    expect(JSON.parse(response.body)).toEqual([{ id: '1', usage: '100MB' }]);
   });
 
-  it('should return 400 if targetId or targetType is missing', async () => {
+  it('should return 400 if userId is missing', async () => {
     const event = {
-      queryStringParameters: {
-        targetId: 'validTargetId',
-      },
+      queryStringParameters: {},
     };
 
     const response = await handler(event);
 
     expect(response.statusCode).toBe(400);
-    expect(JSON.parse(response.body).error).toBe('Please provide both targetId and targetType.');
+    expect(JSON.parse(response.body).error).toBe('Please provide userId');
   });
 
-  it('should return 200 with an empty array if no metadata is found', async () => {
+  it('should return 200 with an empty array if no usage data is found', async () => {
     const event = {
       queryStringParameters: {
-        targetId: 'invalidTargetId',
-        targetType: 'invalidTargetType',
+        userId: 'invalidUserId',
       },
     };
 
@@ -71,8 +66,7 @@ describe('fetchFileMetadata Lambda Function', () => {
 
     const event = {
       queryStringParameters: {
-        targetId: 'validTargetId',
-        targetType: 'validTargetType',
+        userId: 'validUserId',
       },
     };
