+++ backend/storage/src/utils/responseHelper.mjs
