+++ backend/storage/src/utils/s3Utils.mjs
@@ -5,12 +5,12 @@ import {
   DeleteObjectCommand
 } from "@aws-sdk/client-s3";
 import {getSignedUrl} from "@aws-sdk/s3-request-presigner";
-import {createResponse} from '../util/responseHelper.mjs'
+import {createResponse} from './responseHelper.mjs'
 
 const s3 = new S3Client({});
 const BUCKET_NAME = process.env.S3_BUCKET;
 
-export const handler = async (event) => {
+export const requestPresignedUrl = async (event) => {
   console.log(`event is ${JSON.stringify(event, null, 2)}`);
 
   try {
